// typescript/sdk/src/fuel-types/MerkleRootMultisigIsm.ts
export enum ModuleTypeOutput {
  UNUSED = 'UNUSED',
  ROUTING = 'ROUTING',
  AGGREGATION = 'AGGREGATION',
  LEGACY_MULTISIG = 'LEGACY_MULTISIG',
  MERKLE_ROOT_MULTISIG = 'MERKLE_ROOT_MULTISIG',
  MESSAGE_ID_MULTISIG = 'MESSAGE_ID_MULTISIG',
  NULL = 'NULL',
  CCIP_READ = 'CCIP_READ',
}
// ...
const storageSlots: StorageSlot[] = [];

export class MerkleRootMultisigIsmInterface extends Interface {
  constructor() {
    super(abi);
  }

  declare functions: {
    module_type: FunctionFragment;
    verify: FunctionFragment;
    digest: FunctionFragment;
    signature_at: FunctionFragment;
    validators_and_threshold: FunctionFragment;
    initialize: FunctionFragment;
  };
}

export class MerkleRootMultisigIsm extends Contract {
  static readonly abi = abi;
  static readonly storageSlots = storageSlots;

  declare interface: MerkleRootMultisigIsmInterface;
  declare functions: {
    module_type: InvokeFunction<[], ModuleTypeOutput>;
    verify: InvokeFunction<[metadata: Bytes, message: Bytes], boolean>;
    digest: InvokeFunction<[metadata: Bytes, message: Bytes], Bytes>;
    signature_at: InvokeFunction<[metadata: Bytes, index: BigNumberish], Bytes>;
    validators_and_threshold: InvokeFunction<
      [message: Bytes],
      [Vec<EvmAddress>, number]
    >;
    initialize: InvokeFunction<[validators: Vec<EvmAddress>], void>;
  };

  constructor(id: string | Address, accountOrProvider: Account | Provider) {
    super(id, abi, accountOrProvider);
  }
}